Implemented vcache, vcache_hash
and only verified with appending tuples manually. Need to verified with read transactions.
And also need to see if it does not hurt performance.
Current version don't uses O_DIRECT because it seems that the buffer pool of postgres do not as well.
But need to try to use O_DIRECT by adjusting the start address of the mmaped area for vcache blocks.