Fixed several bugs and implemented a naive classification.
Bug fixed 1: Now we do not push down an aborted version into the version cluster anymore, which has xmax value as 0.
Bug fixed 2: Original postgres didn't use a page latch for fetching a tuple from the heap page. Now we are doing in-place update,
we need such synchronization between a reader and an in-place updater. Instead of using page latch which incurs contention overhead,
I've implemented a wait-free logic for handling that race condition.